Save user profile data to store
